Bosses (kinda), shooting damage and premature refactoring! entity.py: Refactoring + introduction of damage to Entities and disappearance of bullets on impact. Changed the temporary colours a bit because I felt like it!
handler.py: Boss HP info, to be made in to a depleting bar like every
other shooting game out there. Introduction of boss entity in to the game
proper.

levels.py: Added a temporary boss enemy to the test script, woo!

scripting.py: Started work on the CreateBoss functionality. Like enemies
and bullets, there will be boss-types. Unlike those guys (or maybe like
future enemies) there may be master-slave boss components in the future.
Example: A boss made up of a body and two arms would have the body as
master and the arms as slaves. Slaves contribute to the total HP which
is held by the body, but may be destroyed separately. If the body is
destroyed, the boss is destroyed. Or the boss might be made up entirely
of slaves. Anyway, right now we're looking at simple single Entity bosses.

Rant time:
I'm starting to wonder why I bothered subclassing Entity in the first place,
especially given how almost every member variable will be computed or set
by script. The remainder (like bounds) could be set by the handler.

That said, occasionally it's useful to have subclassing, such as overriding
collide for bullets. But even then, we could easily have a 'die on any
collide' boolean in Entity.. Maybe I'm overthinking it. Right now, we
have a clear distinction between the different types of Entity, and that's
a pretty good thing, even if their internals aren't so different.

Next up: Player damage, lives and invincibility frames